warnings: fix error and warnings in MPL_DBG_MSG_FMT

+++ src/mpi/coll/iallgather/iallgather_tsp_recexch.c
@@ -26,8 +26,8 @@ static int MPIR_TSP_Iallgather_sched_intra_recexch_data_exchange(int rank, int n
         MPII_Recexchalgo_get_count_and_offset(rank, 0, k, nranks, &count, &offset);
         MPI_Aint send_offset = offset * recv_extent * recvcount;
         MPL_DBG_MSG_FMT(MPIR_DBG_COLL, VERBOSE,
-                        (MPL_DBG_FDEST, "data exchange with %d send_offset %d count %d", partner,
-                         send_offset, count));
+                        (MPL_DBG_FDEST, "data exchange with %d send_offset %ld count %ld",
+                         partner, (long) send_offset, (long) count));
         /* send my data to partner */
         mpi_errno =
             MPIR_TSP_sched_isend(((char *) recvbuf + send_offset), count * recvcount, recvtype,
@@ -39,8 +39,8 @@ static int MPIR_TSP_Iallgather_sched_intra_recexch_data_exchange(int rank, int n
         MPII_Recexchalgo_get_count_and_offset(partner, 0, k, nranks, &count, &offset);
         MPI_Aint recv_offset = offset * recv_extent * recvcount;
         MPL_DBG_MSG_FMT(MPIR_DBG_COLL, VERBOSE,
-                        (MPL_DBG_FDEST, "data exchange with %d recv_offset %d count %d", partner,
-                         recv_offset, count));
+                        (MPL_DBG_FDEST, "data exchange with %d recv_offset %ld count %ld",
+                         partner, (long) recv_offset, (long) count));
         /* recv data from my partner */
         mpi_errno =
             MPIR_TSP_sched_irecv(((char *) recvbuf + recv_offset), count * recvcount, recvtype,
@@ -136,8 +136,8 @@ static int MPIR_TSP_Iallgather_sched_intra_recexch_step2(int step1_sendto, int s
             MPIR_ERR_CHECK(mpi_errno);
             MPL_DBG_MSG_FMT(MPIR_DBG_COLL, VERBOSE,
                             (MPL_DBG_FDEST,
-                             "phase %d nbr is %d send offset %d count %d depend on %d", phase,
-                             nbr, send_offset, count, ((k - 1) * j)));
+                             "phase %d nbr is %d send offset %ld count %ld depend on %d",
+                             phase, nbr, (long) send_offset, (long) count, ((k - 1) * j)));
         }
         /* receive from the neighbors */
         for (i = 0; i < k - 1; i++) {
@@ -158,8 +158,8 @@ static int MPIR_TSP_Iallgather_sched_intra_recexch_step2(int step1_sendto, int s
             nrecvs++;
             MPL_DBG_MSG_FMT(MPIR_DBG_COLL, VERBOSE,
                             (MPL_DBG_FDEST,
-                             "phase %d recv from %d recv offset %d cur_count %d recv returns id[%d] %d",
-                             phase, nbr, recv_offset, count, j * (k - 1) + i,
+                             "phase %d recv from %d recv offset %ld cur_count %ld recv returns id[%d] %d",
+                             phase, nbr, (long) recv_offset, (long) count, j * (k - 1) + i,
                              recv_id[j * (k - 1) + i]));
         }
         if (is_dist_halving == 1)
